+++ org.aion.avm.core/test/org/aion/parallel/TestContract.java
@@ -1,5 +1,6 @@
 package org.aion.parallel;
 
+import java.math.BigInteger;
 import org.aion.avm.api.ABIDecoder;
 import org.aion.avm.api.Address;
 import org.aion.avm.api.BlockchainRuntime;
@@ -18,7 +19,7 @@ public class TestContract {
     }
 
     public static void doTransfer() {
-        BlockchainRuntime.call(deployer, 1000, new byte[0], 100_000L);
+        BlockchainRuntime.call(deployer, BigInteger.valueOf(1000), new byte[0], 100_000L);
     }
 
     public static void addValue() {
